+++ index.d.ts
@@ -24,10 +24,16 @@ export class Expression<T> {
 
   map(array: T, iterator: string, arguments: Record<string, number | T>): T;
   map(array: T, iterator: string, ...arguments: (number | T)[]): T;
+  map(target: T, array: T, iterator: string, arguments: Record<string, number | T>): T;
+  map(target: T, array: T, iterator: string, ...arguments: (number | T)[]): T;
 
   mapAsync(array: T, iterator: string, arguments: Record<string, number | T>): Promise<T>;
   mapAsync(array: T, iterator: string, ...arguments: (number | T)[]): Promise<T>;
   mapAsync(array: T, iterator: string, arguments: Record<string, number | T>, callback: (this: Expression<T>, e: Error | null, r: T | undefined) => void): void;
+
+  mapAsync(target: T, array: T, iterator: string, arguments: Record<string, number | T>): Promise<T>;
+  mapAsync(target: T, array: T, iterator: string, ...arguments: (number | T)[]): Promise<T>;
+  mapAsync(target: T, array: T, iterator: string, arguments: Record<string, number | T>, callback: (this: Expression<T>, e: Error | null, r: T | undefined) => void): void;
 
 
   reduce(array: T, iterator: string, accumulator: string, initializer: number, arguments: Record<string, number | T>): number;